Removed some setter/getter functionality - @property, @[proprety].setter automatically fills/reads back attributes and sometimes doesn't care about internal code (research this). - Removed triton200.py (original) because it was redundant. - Changed triton200p.py -> triton200.py because it was a bit more developed.
